Update Timeline code to latest py3 commit

+++ gnumed/gnumed/client/timelinelib/calendar/num/time.py
@@ -62,12 +62,12 @@ class NumDelta(ComparableValue, GenericDeltaMixin):
 
     def __mul__(self, other):
         # Delta * number
-        return self.__class__(self.value * other)
+        return self.__class__(int(self.value * other))
 
     def __truediv__(self, other):
         if isinstance(other, self.__class__):
             # Delta / Delta
-            return float(self.value) / float(other.value)
+            return self.value / other.value
         else:
             # Delta / number
-            return self.__class__(self.value // other)
+            return self.__class__(int(self.value // other))
